requirements.txt
python-telegram-bot==21.7
anthropic==0.31.0
gspread==6.1.3
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.31.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
